Split audren/audout files into separate folders, some minor cleanup
