--- Ryujinx.Audio/AudioFormat.cs
@@ -1,13 +1,0 @@
-namespace Ryujinx.Audio
-{
-    public enum AudioFormat
-    {
-        Invalid  = 0,
-        PcmInt8  = 1,
-        PcmInt16 = 2,
-        PcmImt24 = 3,
-        PcmImt32 = 4,
-        PcmFloat = 5,
-        Adpcm    = 6
-    }
-}
