+++ Ryujinx.HLE/OsHle/Services/Aud/AudioRenderer/BehaviorIn.cs
@@ -1,6 +1,6 @@
 using System.Runtime.InteropServices;
 
-namespace Ryujinx.HLE.OsHle.Services.Aud
+namespace Ryujinx.HLE.OsHle.Services.Aud.AudioRenderer
 {
     [StructLayout(LayoutKind.Sequential, Size = 0x10, Pack = 4)]
     struct BehaviorIn
