+++ Ryujinx.HLE/OsHle/Services/Aud/AudioRenderer/MemoryPoolContext.cs
@@ -1,4 +1,4 @@
-namespace Ryujinx.HLE.OsHle.Services.Aud
+namespace Ryujinx.HLE.OsHle.Services.Aud.AudioRenderer
 {
     class MemoryPoolContext
     {
