+++ Ryujinx.HLE/OsHle/Services/Aud/AudioRenderer/UpdateDataHeader.cs
@@ -1,4 +1,4 @@
-﻿namespace Ryujinx.HLE.OsHle.Services.Aud
+﻿namespace Ryujinx.HLE.OsHle.Services.Aud.AudioRenderer
 {
     struct UpdateDataHeader
     {
