+++ Ryujinx.HLE/OsHle/Services/Aud/AudioRenderer/WaveBuffer.cs
@@ -1,6 +1,6 @@
 using System.Runtime.InteropServices;
 
-namespace Ryujinx.HLE.OsHle.Services.Aud
+namespace Ryujinx.HLE.OsHle.Services.Aud.AudioRenderer
 {
     [StructLayout(LayoutKind.Sequential, Size = 0x38, Pack = 1)]
     struct WaveBuffer
