+++ Ryujinx.HLE/OsHle/Services/Aud/IAudioRendererManager.cs
@@ -1,6 +1,7 @@
 using Ryujinx.Audio;
 using Ryujinx.HLE.Logging;
 using Ryujinx.HLE.OsHle.Ipc;
+using Ryujinx.HLE.OsHle.Services.Aud.AudioRenderer;
 using System.Collections.Generic;
 
 namespace Ryujinx.HLE.OsHle.Services.Aud
